requirements.txt
certifi==2021.5.30
charset-normalizer==2.0.1
idna==3.2
pypco==1.1.0
python-dotenv==0.18.0
requests==2.26.0
urllib3==1.26.6
